requirements.txt
PyQt6
watchdog
Pillow
pillow-heif
